requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
motor
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
python-dotenv
pymongo